Add integration tests

+++ src/main/java/home/four/you/model/dto/CreateAdResponseDto.java
@@ -21,7 +21,7 @@ public record CreateAdResponseDto(Long id,
                                   Ad.Type type,
                                   String title,
                                   String description,
-                                  BigDecimal price,
+                                  Integer price,
                                   Instant createdAt,
                                   Instant expirationDate,
                                   PropertyDetailsDto property) {
